+++ src/model/points-model.js
@@ -1,5 +1,5 @@
-import { getRandomPoint } from '../mocks/points';
-import { POINTS_COUNT } from '../const';
+import { getRandomPoint } from '../mocks/points.js';
+import { POINTS_COUNT } from '../const.js';
 
 export default class PointsModel {
   #points = Array.from({length: POINTS_COUNT}, getRandomPoint);
